support ts_ad
