+++ paddlex/inference/components/transforms/ts/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .ts_common import *
+from .common import *
