+++ paddlex/inference/models/__init__.py
@@ -14,7 +14,7 @@
 
 
 from pathlib import Path
-from .official_models import official_models
+from ..utils.official_models import official_models
 
 from .base import BasePredictor, BasicPredictor
 from .image_classification import ClasPredictor
@@ -26,6 +26,7 @@ from .instance_segmentation import InstanceSegPredictor
 from .semantic_segmentation import SegPredictor
 from .general_recognition import ShiTuRecPredictor
 from .ts_fc import TSFcPredictor
+from .ts_ad import TSAdPredictor
 from .ts_cls import TSClsPredictor
 from .image_unwarping import WarpPredictor
