+++ paddlex/inference/results/ts.py
@@ -16,6 +16,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
+from ...utils import logging
 from ..utils.io import TSWriter
 from .base import BaseResult
 
@@ -31,30 +32,32 @@ class TSFcResult(BaseResult):
         if not save_path.endswith(".csv"):
             save_path = Path(save_path) / f"{Path(self['ts_path']).stem}.csv"
         self._writer.write(save_path, self["forecast"])
+        logging.info(f"The result has been saved in {save_path}.")
 
 
 class TSClsResult(BaseResult):
 
     def __init__(self, data):
-        super().__init__(
-            {"ts_path": data["ts_path"], "classification": self.process_data(data)}
-        )
+        super().__init__(data)
         self._writer = TSWriter(backend="pandas")
-
-    def process_data(self, data):
-        """apply"""
-        pred_ts = data["forecast"][0]
-        pred_ts -= np.max(pred_ts, axis=-1, keepdims=True)
-        pred_ts = np.exp(pred_ts) / np.sum(np.exp(pred_ts), axis=-1, keepdims=True)
-        classid = np.argmax(pred_ts, axis=-1)
-        pred_score = pred_ts[classid]
-        result = {"classid": [classid], "score": [pred_score]}
-        result = pd.DataFrame.from_dict(result)
-        result.index.name = "sample"
-        return result
 
     def save_to_csv(self, save_path):
         """write ts"""
         if not save_path.endswith(".csv"):
             save_path = Path(save_path) / f"{Path(self['ts_path']).stem}.csv"
         self._writer.write(save_path, self["classification"])
+        logging.info(f"The result has been saved in {save_path}.")
+
+
+class TSAdResult(BaseResult):
+
+    def __init__(self, data):
+        super().__init__(data)
+        self._writer = TSWriter(backend="pandas")
+
+    def save_to_csv(self, save_path):
+        """write ts"""
+        if not save_path.endswith(".csv"):
+            save_path = Path(save_path) / f"{Path(self['ts_path']).stem}.csv"
+        self._writer.write(save_path, self["anomaly"])
+        logging.info(f"The result has been saved in {save_path}.")
